backend/requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
blessed==1.20.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.3
django-picklefield==3.2
django-q2==1.6.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jinxed==1.3.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
mssql-django==1.5
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
pi==0.1.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.8.0
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==3.5.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
setuptools==70.2.0
six==1.16.0
sqlparse==0.5.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13